requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
psycopg2-binary==2.9.10
pyzmq==26.2.1
pykrx==1.2.4
talipp>=2.2.0
polars>=1.0.0
pyarrow>=15.0.0
anthropic>=0.40.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
sentence-transformers>=3.0.0
scikit-learn>=1.5.0
numpy>=1.26.0
sympy>=1.13.0
scipy>=1.14.0
dowhy>=0.11.0
networkx>=3.2.0
fastmcp>=2.0.0
